% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_align.R
\name{adjustRtimePeakGroups}
\alias{adjustRtimePeakGroups}
\title{Auxillary function for adjust R time function}
\usage{
adjustRtimePeakGroups(object, param = PeakGroupsParam())
}
\arguments{
\item{object}{ms object}

\item{param}{xcms::peakgroupsparam object}
}
\value{
pkGrp object
}
\description{
Auxillary function for adjust R time function
}
