% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_align.R
\name{c_do_adjustRtime_peakGroups}
\alias{c_do_adjustRtime_peakGroups}
\title{Title}
\usage{
c_do_adjustRtime_peakGroups(
  peaks,
  peakIndex,
  rtime,
  minFraction = 0.9,
  extraPeaks = 1,
  smooth = c("loess", "linear"),
  span = 0.2,
  family = c("gaussian", "symmetric"),
  peakGroupsMatrix = matrix(ncol = 0, nrow = 0),
  subset = integer(),
  subsetAdjust = c("average", "previous")
)
}
\arguments{
\item{peaks}{The identified peaks and their positions}

\item{peakIndex}{index of peak}

\item{rtime}{retention times of samples}

\item{minFraction}{the minimum number of samples a feature has to exist in to qualify as an "aligning feature"}

\item{extraPeaks}{How many extra peaks are allowed in the area of alignment (can be high for large datasets with high drift)}

\item{smooth}{smoothing of alignment}

\item{span}{how many features should the non-linear alignment model look at simultaniously? close to 0 models highly non-linear drift and close to 1 models linear drift}

\item{family}{kind of model}

\item{peakGroupsMatrix}{the groups (used for minFraction)}

\item{subset}{The code is rewritten such that the subset option is no longer available. All samples will be aligned.}

\item{subsetAdjust}{see "subset"}
}
\value{
adjusted rtimes
}
\description{
Title
}
