% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_filling.R
\name{cfillChromPeaks}
\alias{cfillChromPeaks}
\title{Final fill chrom peak function}
\usage{
cfillChromPeaks(prepared_data = "./tmp/prepared_data.Rdata")
}
\arguments{
\item{prepared_data}{the file/object returned by cXCMS::cPrepFillChromPeaks()}
}
\value{
a preprocessed xcms object
}
\description{
Final fill chrom peak function
}
