% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_filling.R
\name{cFillChromPeaksStep1}
\alias{cFillChromPeaksStep1}
\title{Preparation for the computational intensive peak filling step}
\usage{
cFillChromPeaksStep1(input, output, fcp)
}
\arguments{
\item{input}{path to XCMSset as returned by the peak alignment step}

\item{output}{path of the output file. This file goes into cFillChromPeaksStep2()}

\item{fcp}{the object returned by xcms::FillChromPeaksParam()}
}
\value{
a list object with all the necessary objects for the cGetChromPeakData and cFillChromPeaks functions
}
\description{
cPrepFillChromPeaks, cGetChromPeakData, and cFillChromPeaks are small constructs of fillChromPeaks.

By running the functions as suggested in the vignette, we minimize the RAM usage, thus making the workflow scalable to tens of thousands of files.
}
