% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_filling.R
\name{cFillChromPeaksStep2}
\alias{cFillChromPeaksStep2}
\title{Intensive peak extraction step performed file-wise (parallel in cluster workflows)}
\usage{
cFillChromPeaksStep2(input, output)
}
\arguments{
\item{input}{the file index in sample_info}

\item{output}{the tmp directory for storage of extracted peaks}
}
\value{
a res object for the final peak integration.
}
\description{
Intensive peak extraction step performed file-wise (parallel in cluster workflows)
}
