% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_filling.R
\name{cFillChromPeaksStep3}
\alias{cFillChromPeaksStep3}
\title{fill chrom peak function 3}
\usage{
cFillChromPeaksStep3(inputFromStep1, inputFromStep2, output, param, settings)
}
\arguments{
\item{inputFromStep1}{Output of cFillChromPeaksStep1() function}

\item{inputFromStep2}{Output of cFillChromPeaksStep2() function}

\item{output}{the output which contains the result of the preprocessing}

\item{param}{FillChromPeaksParam object}

\item{settings}{settings list from settings.yaml}
}
\value{
a preprocessed xcms object
}
\description{
fill chrom peak function 3
}
