% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_picking.R
\name{cFindChromPeaks}
\alias{cFindChromPeaks}
\title{Single file peak identification for parallel processing and/or RAM optimization}
\usage{
cFindChromPeaks(sample_info, cwp, index = 1, save_folder = NULL)
}
\arguments{
\item{sample_info}{data.frame (or tibble) containing the columns: "filename" and "group"}

\item{cwp}{native xcms CentWaveParam() function output}

\item{index}{which file should be run in xcms}

\item{save_folder}{the folder for saving the peak picked files. It is recommended only to change this if running custom workflows.}
}
\value{
peak picked files in the save_folder location (./tmp/peaks_identified)
}
\description{
Single file peak identification for parallel processing and/or RAM optimization
}
