% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_picking.R
\name{cFindChromPeaksStep1}
\alias{cFindChromPeaksStep1}
\title{Single file peak identification for parallel processing and/or RAM optimization}
\usage{
cFindChromPeaksStep1(input, output, cwp, metadata = NULL, settings = NULL)
}
\arguments{
\item{input}{filename of the mzML input files}

\item{output}{output filename}

\item{cwp}{native xcms CentWaveParam() function output}

\item{metadata}{sample metadata}

\item{settings}{settings list from settings.yaml}
}
\value{
peak picked files in the save_folder location (./tmp/peaks_identified)
}
\description{
Single file peak identification for parallel processing and/or RAM optimization
}
