% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_picking.R
\name{cFindChromPeaksStep2}
\alias{cFindChromPeaksStep2}
\title{Fast concatenation of XCMS peak picked files}
\usage{
cFindChromPeaksStep2(inputs, output)
}
\arguments{
\item{inputs}{the peak picked files (>1 file names)}

\item{output}{the msnbase element of peak picked files (file name)}
}
\value{
an XCMSset ready for peak grouping and alignment
}
\description{
This function concatenates the files by building a string followed by concatenation.
By doing it this way we reduce running time from n^2 to n allowing us to concatenate several thousand files in minutes rather than days.
}
