% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_picking.R
\name{collect_files}
\alias{collect_files}
\title{Fast concatenation of XCMS peak picked files}
\usage{
collect_files(peaks_identified_path = "./tmp/peaks_identified/")
}
\arguments{
\item{peaks_identified_path}{the location of the peak picked files}
}
\value{
an XCMSset ready for peak grouping and alignment
}
\description{
Fast concatenation of XCMS peak picked files
}
