% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_filling.R
\name{cPrepFillChromPeaks}
\alias{cPrepFillChromPeaks}
\title{Preparation for the computational intensive peak filling step}
\usage{
cPrepFillChromPeaks(object, param, msLevel = 1L, output_dir = "./tmp")
}
\arguments{
\item{object}{XCMSset as returned by the peak alignment step}

\item{param}{the object returned by xcms::FillChromPeaksParam()}

\item{msLevel}{level 1 or 2. Default 1 for untargeted metabolomics}
}
\value{
a list object with all the necessary objects for the cGetChromPeakData and cFillChromPeaks functions
}
\description{
cPrepFillChromPeaks, cGetChromPeakData, and cFillChromPeaks are small constructs of fillChromPeaks.

By running the functions as suggested in the vignette, we minimize the RAM usage, thus making the workflow scalable to tens of thousands of files.
}
