% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_picking.R
\name{.checkdir}
\alias{.checkdir}
\title{Create output dir}
\usage{
.checkdir(output)
}
\arguments{
\item{output}{the output path}
}
\value{
None
}
\description{
This function is a utility function that creates the output directory if it does not already exist
}
