% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_filling.R
\name{[}
\alias{[}
\alias{[,XcmsExperiment,ANY,ANY,ANY-method}
\title{Subset XcmsExperiment}
\usage{
\S4method{[}{XcmsExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{XcmsExperiment object}

\item{i}{row indices}

\item{j}{column indices}

\item{...}{additional arguments}

\item{drop}{logical, whether to drop dimensions}
}
\value{
subsetted XcmsExperiment
}
\description{
Subset XcmsExperiment
}
\keyword{internal}
