% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_align.R
\name{adjustRtime}
\alias{adjustRtime}
\title{Adjust retention time}
\usage{
adjustRtime(object, param)
}
\arguments{
\item{object}{ms object}

\item{param}{xcms::PeakGroupsParam}
}
\value{
aligned ms object
}
\description{
Adjust retention time
}
