% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_filling.R
\name{cFillChromPeaksStep1}
\alias{cFillChromPeaksStep1}
\title{Preparation for the computational intensive peak filling step}
\usage{
cFillChromPeaksStep1(input, output, param, interval, settings)
}
\arguments{
\item{input}{path to XcmsExperiment as returned by the peak alignment step}

\item{output}{path(s) of the output file(s). This file goes into cFillChromPeaksStep2()}

\item{param}{the object returned by xcms::FillChromPeaksParam()}

\item{interval}{numeric vector of length 2 with start and end sample indices}

\item{settings}{settings list from settings.yaml}
}
\value{
a list object with all the necessary objects for the cGetChromPeakData and cFillChromPeaks functions
}
\description{
cPrepFillChromPeaks, cGetChromPeakData, and cFillChromPeaks are small constructs of fillChromPeaks.

By running the functions as suggested in the vignette, we minimize the RAM usage, thus making the workflow scalable to tens of thousands of files.
}
