% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_filling.R
\name{cFillChromPeaksStep2}
\alias{cFillChromPeaksStep2}
\title{Intensive peak extraction step performed sample-wise}
\usage{
cFillChromPeaksStep2(input, output, index)
}
\arguments{
\item{input}{the file index in sample_info}

\item{output}{the tmp directory for storage of extracted peaks}

\item{index}{the index of the sample in the sample list (i.e., first file=1, second file=2...)}
}
\value{
a res object for the final peak integration.
}
\description{
This function is the second step in the peak filling process. It is performed file-wise and in parallel.
}
