% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_filling.R
\name{cFillChromPeaksStep2}
\alias{cFillChromPeaksStep2}
\title{Intensive peak extraction step performed file-wise (parallel in cluster workflows)}
\usage{
cFillChromPeaksStep2(input, output)
}
\arguments{
\item{input}{path to the prepared data file from cFillChromPeaksStep1}

\item{output}{the output file path for storage of extracted peaks}
}
\value{
NULL (writes output to file)
}
\description{
Intensive peak extraction step performed file-wise (parallel in cluster workflows)
}
