% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_filling.R
\name{cFillChromPeaksStep3}
\alias{cFillChromPeaksStep3}
\title{fill chrom peak function 3}
\usage{
cFillChromPeaksStep3(inputFromStep1, inputFromStep2, output)
}
\arguments{
\item{inputFromStep1}{Output of cFillChromPeaksStep1() function}

\item{inputFromStep2}{Output of cFillChromPeaksStep2() function}

\item{output}{the output which contains the result of the preprocessing}
}
\value{
a preprocessed xcms object
}
\description{
This function is the third step in the peak filling process.
It collects the files and generates the final integrated output (intragation of missing values)
}
