% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_picking.R
\name{cFindChromPeaksStep1}
\alias{cFindChromPeaksStep1}
\title{Single file peak identification for parallel processing and/or RAM optimization}
\usage{
cFindChromPeaksStep1(input, output, cwp, groups = NULL)
}
\arguments{
\item{input}{filename of the mzML input files}

\item{output}{output filename}

\item{cwp}{native xcms CentWaveParam() function output}

\item{groups}{the group variable used in new("NAnnotatedDataFrame", data.frame("sample_name"=sample_name, "group"=groups)). The group variable might bias the analysis and in big datasets it might beneficial setting groups = None to do random group assignment}
}
\value{
peak picked files in the save_folder location (./tmp/peaks_identified)
}
\description{
Single file peak identification for parallel processing and/or RAM optimization
}
