% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_picking.R
\name{cFindChromPeaksStep2}
\alias{cFindChromPeaksStep2}
\title{Fast concatenation of XCMS peak picked files}
\usage{
cFindChromPeaksStep2(inputs, output)
}
\arguments{
\item{inputs}{the peak picked files}

\item{output}{the msnbase element of peak picked files}
}
\value{
an XCMSset ready for peak grouping and alignment
}
\description{
Fast concatenation of XCMS peak picked files
}
